requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
mysqlclient==2.2.0